Optimisations around `store`-`store-view` interaction
Now for big stores view works faster also for the records at postions
 above 25k-50k.

`DataStoreBase::get()` was returning postion which required linear
iteration in over data store. For indexes which were far away from
beginning of the store collection it was major overhead.

Now position is a function of the store view so in worst case you need
to check page size of data in the view which is expected to be smaller
then what there is in the store.