Id allocator for record is now defined in record
Until now we've been passing the id allocator for record as type
argument. In long run this ended a super painfull and error prone
(since we've suplied defaults). Now it's in the record itself. What
needs to access it knows the record type, so it's ok (sometimes just
turbo fish school, but there is nothing we can do about that).