Make backend the backend
Until now we were implementing backends as `DataStore` instances. This
was not bad since it allowed to just slap the backend implementation
wherever you would like to have a store. Downside of it was
`Rc<RefCell<Backend>>`. This made user of the library track down the
borrows for the backend inside of their code. It shoudn't be like that.

 So as part of cleaning up `StoreIdAllocators` the `Store` structure was
 introduced. Commit ago it was just a wrapper for the backend which
removed the `Rc-RefCell` hell from the user.

This commit takes it to another level. From now Store will be the only
`DataStore` implementation provided. It also handles both subscriptions
of the views and sotre message handling.

Benefits for this split are

1. Backend implementation is much simpler and is less affected by the
`Rc-RefCell` hell.
2. Testing of the backends is much easier
3. User code is still clean

First 2 points are critical for us. Users is somewhat expected to write
their own backend. We can't solve all their requirements around data
storage. Easier it is for user to create their custom backend the
better. As long as backend works as expected rest of the stack can be
in most cases provided.